Make compatible with dev15 RC
The VSIXInstaller.exe now requires both TargetVersion and Target SKU for installation.
Since there is no built-in way to detect the installed SKU, we just attempt to install
to all SKUs. The ones not installed will silently fail.

Note that we also install to each IDE with its own VSIXInstaller.exe, to avoid potential
incompatibility issues.